增加了根据id查询一条信息，根据code查询出最新的一条信息两个接口

+++ blade-common/src/main/java/org/springblade/common/entity/Demand.java
@@ -186,5 +186,5 @@ public class Demand implements Serializable {
     private GoodsTypeEntity goodsTypeEntity;
 
     @TableField(exist = false)
-    private int quotationNumber;
+    private Long quotationNumber;
 }
